requirements.txt
torch
git+https://github.com/huggingface/transformers.git
gradio
Pillow
pypdfium2
spaces
openai
jinja2
vllm
pandas
openpyxl